Update clamav impl, add Thunderbird support
Adds some finicky support for scanning Thunderbird attachments. This
required wrapping Thunderbird in a script which overrides the tmp
directory it tries to use (when opening attachments). We set that to a
custom tmp directory instead (/tmp/thunderbird) such that clamonacc only
scans that directory (instead of everything in /tmp).

I also had an issue where clamonacc would use 100% CPU here. I had to
clear out the quarantine directory to fix this, which is unfortunate.